LPS-145191 Extract managed service factory to a different class
